+++ lib/src/core/comp.c_input_border.dart
@@ -256,5 +256,5 @@ class COutlineInputBorder extends InputBorder {
   }
 
   @override
-  int get hashCode => hashValues(borderSide, borderRadius, gapPadding);
+  int get hashCode => Object.hash(borderSide, borderRadius, gapPadding);
 }