+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.7
+
+Added ListTile.radio and CRadioButton
+
+## 0.0.6
+
+Fix bugs related to borderRadius of divided linear progress indicator
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+- TODO: Describe initial release.